Make the small screen banner consistent
The regular banner now looks like the small banner at the smallest
screen size, this is because the link needs to be visible for legal
reasons in a lot of cases, and gives us more room.

Currently there's a bit of copy/paste in here, but I didn't want to use
themed mixins inside the non-themed ones. Open to ideas if there's a
nicer way of doing this.

This resolves #40