Allow for empty or partial banner content
Now when you provide an empty banner element or a banner element with
only partial markup, we construct the rest of the markup around it.

With an empty element we construct the banner inside it, using data
attributes to configure it:

```html
<div
    class="o-banner"
    data-o-component="o-banner"
    data-o-banner-content-long="Do you like o-banner?"
    data-o-banner-button-label="Yes"
></div>
```

When the element has content but does not include an `o-banner__outer`
element, then we wrap the given content with the appropriate markup:

```html
<div class="o-banner" data-o-component="o-banner">
    <p>This is the banner content</p>
</div>
```

This stuff is needed to make cookie-message a little neater and easier
to understand, as we'll be constructing it from an empty element in a
lot of cases.