Split themes into themes and layouts. Upgrade dependencies.
Two layouts may not be used together and two themes may not be used
together, so this commit split them into two options.

Unfortunately these changes came as part of upgrading dependencies,
so this commit is bigger than it should be. It uses the breaking
branch for dependencies until stable betas are released.

Note: Button alignment is incorrect. This is a problem with o-buttons.
Note 2: Button text colour is different, it no longer matches the
colour of the banner background. If this is what we want it should be
set by o-buttons. I'll bring this up with UXD and make any required
changes.